# museum-service/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: team@redocly.com
    url: https://redocly.com/docs/cli/
  description: "Imaginary, but delightful Museum API for interacting with museum services\
    \ and information. Built with love by Redocly."
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  termsOfService: https://redocly.com/subscription-agreement/
  title: Redocly Museum API
  version: 1.1.0
servers:
- url: http://localhost:8080/v1
security:
- MuseumPlaceholderAuth: []
...
paths:
  /museum-hours:
    get:
      description: Get upcoming museum operating hours.
      operationId: getMuseumHours
      parameters:
      - $ref: '#/components/parameters/StartDate'
      - $ref: '#/components/parameters/PaginationPage'
      - $ref: '#/components/parameters/PaginationLimit'
      responses:
        "200":
          content:
            application/json:
              examples:
                default_example:
                  $ref: '#/components/examples/GetMuseumHoursResponseExample'
              schema:
                $ref: '#/components/schemas/GetMuseumHoursResponse'
          description: Success.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Get museum hours
      tags:
      - Operations
      x-accepts: "application/json,application/problem+json"
      x-tags:
      - tag: Operations
...
  /tickets:
    post:
      description: Purchase museum tickets for general entry or special events.
      operationId: buyMuseumTickets
      requestBody:
        content:
          application/json:
            examples:
              general_entry:
                $ref: '#/components/examples/BuyGeneralTicketsRequestExample'
              event_entry:
                $ref: '#/components/examples/BuyEventTicketsRequestExample'
            schema:
              $ref: '#/components/schemas/BuyMuseumTicketsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                general_entry:
                  $ref: '#/components/examples/BuyGeneralTicketsResponseExample'
                event_entry:
                  $ref: '#/components/examples/BuyEventTicketsResponseExample'
              schema:
                $ref: '#/components/schemas/BuyMuseumTicketsResponse'
          description: Success.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Buy museum tickets
      tags:
      - Tickets
      x-content-type: application/json
      x-accepts: "application/json,application/problem+json"
      x-tags:
      - tag: Tickets
  /tickets/{ticketId}/qr:
    get:
      description: Return an image of your ticket with scannable QR code. Used for
        event entry.
      operationId: getTicketCode
      parameters:
      - $ref: '#/components/parameters/TicketId'
      responses:
        "200":
          content:
            image/png:
              schema:
                $ref: '#/components/schemas/GetTicketCodeResponse'
          description: Scannable event ticket in image format.
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
      summary: Get ticket QR code
      tags:
      - Tickets
      x-accepts: "application/problem+json,image/png"
      x-tags:
      - tag: Tickets
components:
  examples:
    BuyGeneralTicketsRequestExample:
      summary: General entry ticket
      value:
        ticketType: general
        ticketDate: 2023-09-07
        email: todd@example.com
    BuyEventTicketsRequestExample:
      summary: Special event ticket
      value:
        ticketType: general
        eventId: dad4bce8-f5cb-4078-a211-995864315e39
        ticketDate: 2023-09-05
        email: todd@example.com
    BuyGeneralTicketsResponseExample:
      summary: General entry ticket
      value:
        message: Museum general entry ticket purchased
        ticketId: 382c0820-0530-4f4b-99af-13811ad0f17a
        ticketType: general
        ticketDate: 2023-09-07
        confirmationCode: ticket-general-e5e5c6-dce78
    BuyEventTicketsResponseExample:
      summary: Special event ticket
      value:
        message: Museum special event ticket purchased
        ticketId: b811f723-17b2-44f7-8952-24b03e43d8a9
        eventName: Mermaid Treasure Identification and Analysis
        ticketType: event
        ticketDate: 2023-09-05
        confirmationCode: ticket-event-9c55eg-8v82a
...
    GetMuseumHoursResponseExample:
      summary: Get hours response
      value:
      - date: 2023-09-11
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-12
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-13
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-14
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-15
        timeOpen: 10:00
        timeClose: 16:00
      - date: 2023-09-18
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-19
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-20
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-21
        timeOpen: 09:00
        timeClose: 18:00
      - date: 2023-09-22
        timeOpen: 10:00
        timeClose: 16:00
  parameters:
    PaginationPage:
      description: Page number to retrieve.
      explode: true
      in: query
      name: page
      required: false
      schema:
        default: 1
        example: 2
        type: integer
      style: form
    PaginationLimit:
      description: Number of days per page.
      explode: true
      in: query
      name: limit
      required: false
      schema:
        default: 10
        example: 15
        maximum: 30
        type: integer
      style: form
...
    StartDate:
      description: Starting date to retrieve future operating hours from. Defaults
        to today's date.
      explode: true
      in: query
      name: startDate
      required: false
      schema:
        example: 2023-02-23
        format: date
        type: string
      style: form
...
    TicketId:
      description: Identifier for a ticket to a museum event. Used to generate ticket
        image.
      explode: false
      in: path
      name: ticketId
      required: true
      schema:
        example: a54a57ca-36f8-421b-a6b4-2e8f26858a4c
        format: uuid
        type: string
      style: simple
  responses:
    BadRequest:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Bad request.
    NotFound:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Not found.
...
  schemas:
    TicketType:
      description: Type of ticket being purchased. Use `general` for regular museum
        entry and `event` for tickets to special events.
      enum:
      - event
      - general
      example: event
      type: string
...
    BuyMuseumTicketsRequest:
      description: Request payload used for purchasing museum tickets.
      example:
        eventId: 3be6453c-03eb-4357-ae5a-984a0e574a54
        phone: +1(234)-567-8910
        ticketType: event
        ticketDate: 2023-10-29
        email: museum-lover@example.com
      properties:
        ticketType:
          $ref: '#/components/schemas/TicketType'
        eventId:
          description: Identifier for a special event.
          example: 3be6453c-03eb-4357-ae5a-984a0e574a54
          format: uuid
          type: string
        ticketDate:
          example: 2023-10-29
          format: date
          type: string
        email:
          description: Email address for ticket purchaser.
          example: museum-lover@example.com
          format: email
          type: string
        phone:
          description: Phone number for the ticket purchaser (optional).
          example: +1(234)-567-8910
          type: string
      required:
      - email
      - ticketDate
      - ticketType
...
    BuyMuseumTicketsResponse:
      description: Details for a museum ticket after a successful purchase.
      example:
        confirmationCode: ticket-event-a98c8f-7eb12
        eventName: Pirate Coding Workshop
        ticketType: event
        message: Museum general entry ticket purchased
        ticketId: a54a57ca-36f8-421b-a6b4-2e8f26858a4c
        ticketDate: 2023-10-29
      properties:
        message:
          description: Confirmation message after a ticket purchase.
          example: Museum general entry ticket purchased
          type: string
        eventName:
          description: Name of the special event.
          example: Pirate Coding Workshop
          type: string
        ticketId:
          description: Unique identifier for museum ticket. Generated when purchased.
          example: a54a57ca-36f8-421b-a6b4-2e8f26858a4c
          format: uuid
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        ticketDate:
          example: 2023-10-29
          format: date
          type: string
        confirmationCode:
          description: Unique confirmation code used to verify ticket purchase.
          example: ticket-event-a98c8f-7eb12
          type: string
      required:
      - confirmationCode
      - message
      - ticketDate
      - ticketId
      - ticketType
    GetTicketCodeResponse:
      description: Image of a ticket with a QR code used for museum or event entry.
      format: binary
      type: string
    GetMuseumHoursResponse:
      description: List of museum operating hours for consecutive days.
      items:
        $ref: '#/components/schemas/MuseumDailyHours'
      type: array
    MuseumDailyHours:
      description: Daily operating hours for the museum.
      example:
        date: 2023-10-29
        timeClose: 18:00
        timeOpen: 09:00
      properties:
        date:
          example: 2023-10-29
          format: date
          type: string
        timeOpen:
          description: Time the museum opens on a specific date. Uses 24 hour time
            format (`HH:mm`).
          example: 09:00
          pattern: "^([01]\\d|2[0-3]):?([0-5]\\d)$"
          type: string
        timeClose:
          description: Time the museum closes on a specific date. Uses 24 hour time
            format (`HH:mm`).
          example: 18:00
          pattern: "^([01]\\d|2[0-3]):?([0-5]\\d)$"
          type: string
      required:
      - date
      - timeClose
      - timeOpen
...
    Error:
      properties:
        type:
          example: object
          type: string
        title:
          example: Validation failed
          type: string
  securitySchemes:
    MuseumPlaceholderAuth:
      scheme: basic
      type: http